+++ integration_tests/src/test/kotlin/utils/SystemProperty.kt
@@ -1,0 +1,10 @@
+package utils
+
+import java.util.Properties
+
+fun Map<String, String>.getPropertyAsList(key: String, default: List<String>) = get(key).orEmpty().run {
+    if (isBlank()) default
+    else split(",")
+}
+
+fun Properties.toStringMap() = map { Pair(it.key.toString(), it.value.toString()) }.toMap()