+++ test_runner/src/main/kotlin/ftl/run/RefreshLastRun.kt
@@ -15,6 +15,7 @@ import ftl.run.common.updateMatrixFile
 import ftl.args.ShardChunks
 import ftl.json.needsUpdate
 import ftl.json.updateWithMatrix
+import ftl.json.validate
 import ftl.util.MatrixState
 import kotlinx.coroutines.Deferred
 import kotlinx.coroutines.Dispatchers
@@ -33,6 +34,8 @@ suspend fun refreshLastRun(currentArgs: IArgs, testShardChunks: ShardChunks) {
 
     // Must generate reports *after* fetching xml artifacts since reports require xml
     ReportManager.generate(matrixMap, lastArgs, testShardChunks)
+
+    matrixMap.validate(lastArgs.ignoreFailedTests)
 }
 
 /** Refresh all in progress matrices in parallel **/