+++ test_runner/src/main/kotlin/ftl/run/platform/android/CreateInstrumentationConfig.kt
@@ -2,6 +2,7 @@ package ftl.run.platform.android
 
 import ftl.args.AndroidArgs
 import ftl.run.model.InstrumentationTestContext
+import ftl.shard.testCases
 
 internal fun AndroidArgs.createInstrumentationConfig(
     testApk: InstrumentationTestContext
@@ -12,6 +13,6 @@ internal fun AndroidArgs.createInstrumentationConfig(
     orchestratorOption = "USE_ORCHESTRATOR".takeIf { useOrchestrator },
     disableSharding = disableSharding,
     numUniformShards = numUniformShards,
-    testShards = testApk.shards,
+    testShards = testApk.shards.testCases,
     keepTestTargetsEmpty = disableSharding && testTargets.isEmpty()
 )