+++ test_runner/src/test/kotlin/ftl/args/ArgsHelperTest.kt
@@ -10,6 +10,8 @@ import ftl.args.yml.mergeYmlKeys
 import ftl.config.FtlConstants
 import ftl.config.common.CommonGcloudConfig
 import ftl.config.ios.IosGcloudConfig
+import ftl.gc.GcStorage
+import ftl.gc.GcStorage.exist
 import ftl.shard.TestMethod
 import ftl.shard.TestShard
 import ftl.shard.stringShards
@@ -19,6 +21,7 @@ import ftl.test.util.assertThrowsWithMessage
 import ftl.run.exception.FlankGeneralError
 import ftl.run.exception.FlankConfigurationError
 import io.mockk.every
+import io.mockk.mockkObject
 import io.mockk.spyk
 import io.mockk.unmockkAll
 import org.junit.After
@@ -193,7 +196,10 @@ class ArgsHelperTest {
 
         val args = spyk(AndroidArgs.default())
         every { args.maxTestShards } returns maxTestShards
-        assertThrowsWithMessage(Throwable::class, "max-test-shards must be >= ${IArgs.AVAILABLE_PHYSICAL_SHARD_COUNT_RANGE.first} and <= ${IArgs.AVAILABLE_PHYSICAL_SHARD_COUNT_RANGE.last}. But current is $maxTestShards") {
+        assertThrowsWithMessage(
+            Throwable::class,
+            "max-test-shards must be >= ${IArgs.AVAILABLE_PHYSICAL_SHARD_COUNT_RANGE.first} and <= ${IArgs.AVAILABLE_PHYSICAL_SHARD_COUNT_RANGE.last}. But current is $maxTestShards"
+        ) {
             assertCommonProps(args)
         }
     }
@@ -213,6 +219,24 @@ class ArgsHelperTest {
         every { args.repeatTests } returns 0
         assertThrowsWithMessage(Throwable::class, "num-test-runs must be >= 1") {
             assertCommonProps(args)
+        }
+    }
+
+    @Test
+    fun `should throw an error if apk file does not exists on gcs`() {
+        mockkObject(GcStorage) {
+            every { exist("gs://any-bucket/any-file.apk") } returns false
+            assertThrowsWithMessage(FlankGeneralError::class, "'gs://any-bucket/any-file.apk' from apk doesn't exist") {
+                assertFileExists("gs://any-bucket/any-file.apk", "from apk")
+            }
+        }
+    }
+
+    @Test
+    fun `should not throw and error if apk file exist on gcs`() {
+        mockkObject(GcStorage) {
+            every { exist("gs://any-bucket/any-file.apk") } returns true
+            assertFileExists("gs://any-bucket/any-file.apk", "from apk")
         }
     }
 }