+++ test_runner/src/test/kotlin/ftl/reports/utils/ReportManagerTest.kt
@@ -3,6 +3,7 @@ package ftl.reports.utils
 import com.google.common.truth.Truth.assertThat
 import ftl.args.AndroidArgs
 import ftl.gc.GcStorage
+import ftl.json.validate
 import ftl.reports.CostReport
 import ftl.reports.FullJUnitReport
 import ftl.reports.JUnitReport
@@ -52,6 +53,7 @@ class ReportManagerTest {
         every { mockArgs.smartFlankGcsPath } returns ""
         every { mockArgs.useLegacyJUnitResult } returns true
         ReportManager.generate(matrix, mockArgs, emptyList())
+        matrix.validate()
     }
 
     @Test
@@ -229,8 +231,13 @@ class ReportManagerTest {
     fun `should get weblink from legacy path and ios path`() {
         val legacyPath = File("results/2020-08-06_12-08-55.641213_jGpY/matrix_0/NexusLowRes-28-en-portrait/test_result_1.xml")
         val iosPath = File("results/test_dir/shard_0/iphone8-12.0-en-portrait/test_result_0.xml")
-        val objectName = "object_name"
-        assertEquals("object_name/2020-08-06_12-08-55.641213_jGpY/matrix_0", legacyPath.getMatrixPath(objectName))
-        assertEquals("object_name/test_dir/shard_0", iosPath.getMatrixPath(objectName))
+        assertEquals("2020-08-06_12-08-55.641213_jGpY/matrix_0", legacyPath.getMatrixPath("2020-08-06_12-08-55.641213_jGpY"))
+        assertEquals("test_dir/shard_0", iosPath.getMatrixPath("test_dir"))
+    }
+
+    @Test
+    fun `shouldn't contains multiple test_dir in MatrixPath`() {
+        val path = File("results/test_dir/test_dir/shard_0/iphone8-12.0-en-portrait/test_result_0.xml")
+        assertEquals("test_dir/shard_0", path.getMatrixPath("test_dir"))
     }
 }